chore: move all unused code to unused/experiments.cpp
